Rename plotting functions, add __inits__ of new submodules, add aliases for old names
