+++ src/crested/pl/corr/__init__.py
@@ -1,0 +1,5 @@
+"""Initialize the corr module."""
+
+
+from ._heatmap import heatmap_predictions, heatmap_self
+from ._scatter import scatter
