+++ src/crested/pl/region/__init__.py
@@ -1,0 +1,5 @@
+"""Initialize the region module."""
+
+from ._bar import bar
+from ._scatter import scatter
+
